# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planificacion Horaria",
        "description": "Gestion de horarios de clases",
        "contact": {
            "email": "contacto@tudominio.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://www.terciarioUrquiza.com.ar/planificacionHoraria"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/materias": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Obtener todas las materias",
                "description": "Devuelve un array de materias",
                "operationId": "ce64d4a8c3503884806e1c48739e20fc",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "500": {
                        "description": "Error al obtener las materias"
                    }
                }
            }
        },
        "/api/materias/{id}": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Obtener materia por id",
                "description": "Devuelve una materia",
                "operationId": "8624bdc19c63b1ac3eb44f7d29e24eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró la materia"
                    },
                    "500": {
                        "description": "Error al obtener la materia"
                    }
                }
            }
        },
        "/api/materias/guardar": {
            "post": {
                "tags": [
                    "Materias"
                ],
                "summary": "Guardar materia",
                "description": "Guarda una materia",
                "operationId": "b59f92a268f6c4a79a80cc61499965fa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "modulos_semanales": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa"
                    },
                    "500": {
                        "description": "Error al guardar la materia"
                    }
                }
            }
        },
        "/api/materias/actualizar/{id}": {
            "put": {
                "tags": [
                    "Materias"
                ],
                "summary": "Actualizar materia",
                "description": "Actualiza una materia",
                "operationId": "cdf37fed87835a0a03d131e9019d8b30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "modulos_semanales": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró la materia"
                    },
                    "500": {
                        "description": "Error al actualizar la materia"
                    }
                }
            }
        },
        "/api/materias/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Materias"
                ],
                "summary": "Eliminar materia",
                "description": "Elimina una materia",
                "operationId": "a26f40f950dd4798fd21783450374f78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró la materia"
                    },
                    "500": {
                        "description": "Error al eliminar la materia"
                    }
                }
            }
        }
    },
...
}